+++ typescript/cli/ci-test.sh
@@ -38,8 +38,8 @@ yarn workspace @hyperlane-xyz/cli run hyperlane deploy core \
     --targets anvil1,anvil2 \
     --chains ./examples/anvil-chains.yaml \
     --artifacts /tmp/empty-artifacts.json \
-    --ism ./examples/multisig-ism.yaml \
-    --hook ./examples/hook-config.yaml \
+    --ism ./examples/ism.yaml \
+    --hook ./examples/hooks.yaml \
     --out /tmp \
     --key 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80 \
     --yes