+++ typescript/infra/config/environments/mainnet3/core.ts
@@ -36,8 +36,8 @@ export const core: ChainMap<CoreConfig> = objMap(owners, (local, owner) => {
 
   const requiredHook: ProtocolFeeHookConfig = {
     type: HookType.PROTOCOL_FEE,
-    maxProtocolFee: ethers.utils.parseUnits('1', 'gwei'), // 1 gwei of native token
-    protocolFee: BigNumber.from(0), // 0 wei
+    maxProtocolFee: ethers.utils.parseUnits('1', 'gwei').toString(), // 1 gwei of native token
+    protocolFee: BigNumber.from(0).toString(), // 0 wei
     beneficiary: owner,
     owner,
   };