+++ typescript/infra/config/environments/testnet4/helloworld.ts
@@ -13,7 +13,7 @@ export const hyperlaneHelloworld: HelloWorldConfig = {
   kathy: {
     docker: {
       repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-      tag: '52a8416-20231024-025635',
+      tag: 'e152268-20231205-122327',
     },
     chainsToSkip: [],
     runEnv: environment,
@@ -33,7 +33,7 @@ export const releaseCandidateHelloworld: HelloWorldConfig = {
   kathy: {
     docker: {
       repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-      tag: '52a8416-20231024-025635',
+      tag: 'e152268-20231205-122327',
     },
     chainsToSkip: [],
     runEnv: environment,