+++ typescript/infra/src/config/agent/relayer.ts
@@ -16,7 +16,12 @@ import { AgentAwsUser } from '../../agents/aws';
 import { Role } from '../../roles';
 import { HelmStatefulSetValues } from '../infrastructure';
 
-import { AgentConfigHelper, KeyConfig, RootAgentConfig } from './agent';
+import {
+  AgentConfigHelper,
+  KeyConfig,
+  RootAgentConfig,
+  defaultChainSignerKeyConfig,
+} from './agent';
 
 export { GasPaymentEnforcement as GasPaymentEnforcementConfig } from '@hyperlane-xyz/sdk';
 
@@ -60,7 +65,7 @@ export class RelayerConfigHelper extends AgentConfigHelper<RelayerConfig> {
     const baseConfig = this.#relayerConfig!;
 
     const relayerConfig: RelayerConfig = {
-      relayChains: this.contextChainNames[Role.Relayer].join(','),
+      relayChains: this.relayChains.join(','),
       gasPaymentEnforcement: JSON.stringify(baseConfig.gasPaymentEnforcement),
     };
 
@@ -84,6 +89,8 @@ export class RelayerConfigHelper extends AgentConfigHelper<RelayerConfig> {
 
   // Get the signer configuration for each chain by the chain name.
   async signers(): Promise<ChainMap<KeyConfig>> {
+    let chainSigners: ChainMap<KeyConfig> = {};
+
     if (this.aws) {
       const awsUser = new AgentAwsUser(
         this.runEnv,
@@ -93,25 +100,24 @@ export class RelayerConfigHelper extends AgentConfigHelper<RelayerConfig> {
       );
       await awsUser.createIfNotExists();
       const awsKey = (await awsUser.createKeyIfNotExists(this)).keyConfig;
-      return Object.fromEntries(
-        this.contextChainNames[Role.Relayer].map((name) => {
-          const chain = chainMetadata[name];
-          // Sealevel chains always use hex keys
-          if (chain?.protocol == ProtocolType.Sealevel) {
-            return [name, { type: AgentSignerKeyType.Hex }];
-          } else {
-            return [name, awsKey];
-          }
-        }),
-      );
-    } else {
-      return Object.fromEntries(
-        this.contextChainNames[Role.Relayer].map((name) => [
-          name,
-          { type: AgentSignerKeyType.Hex },
-        ]),
-      );
+
+      // AWS keys only work for Ethereum chains
+      for (const chainName of this.relayChains) {
+        if (chainMetadata[chainName].protocol === ProtocolType.Ethereum) {
+          chainSigners[chainName] = awsKey;
+        }
+      }
     }
+
+    // For any chains that were not configured with AWS keys, fill in the defaults
+    for (const chainName of this.relayChains) {
+      if (chainSigners[chainName] !== undefined) {
+        continue;
+      }
+      chainSigners[chainName] = defaultChainSignerKeyConfig(chainName);
+    }
+
+    return chainSigners;
   }
 
   // Returns whether the relayer requires AWS credentials
@@ -129,6 +135,10 @@ export class RelayerConfigHelper extends AgentConfigHelper<RelayerConfig> {
 
   get role(): Role {
     return Role.Relayer;
+  }
+
+  get relayChains(): Array<string> {
+    return this.contextChainNames[Role.Relayer];
   }
 }