+++ solidity/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @hyperlane-xyz/core
 
+## 3.3.0
+
+### Patch Changes
+
+- 350175581: Rename StaticProtocolFee hook to ProtocolFee for clarity
+  - @hyperlane-xyz/utils@3.3.0
+
+## 3.2.0
+
+### Minor Changes
+
+- df34198d4: Includes storage gap in Mailbox Client for forwards compatibility
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@3.2.0
+
 ## 3.1.10
 
 ### Patch Changes