+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @hyperlane-xyz/helloworld
 
+## 3.3.0
+
+### Patch Changes
+
+- Updated dependencies [7e620c9df]
+- Updated dependencies [350175581]
+- Updated dependencies [9f2c7ce7c]
+  - @hyperlane-xyz/sdk@3.3.0
+  - @hyperlane-xyz/core@3.3.0
+
+## 3.2.0
+
+### Patch Changes
+
+- Updated dependencies [df34198d4]
+- Updated dependencies [df693708b]
+  - @hyperlane-xyz/core@3.2.0
+  - @hyperlane-xyz/sdk@3.2.0
+
 ## 3.1.10
 
 ### Patch Changes